add class TransverseMercator
replace Grid form/to LatLon transformation method of PGTaskMgr class
