+++ Common/Source/Calc/Task/PGTask/PGTaskMgr.h
@@ -15,7 +15,9 @@
 #include "PGTaskPt.h"
 #include "vector"
 
-class PGTaskMgr {
+class TransverseMercator;
+
+class PGTaskMgr final {
 public:
     PGTaskMgr();
     virtual ~PGTaskMgr();
@@ -23,12 +25,14 @@ public:
     void Initialize();
     void Optimize(NMEA_INFO *Basic, DERIVED_INFO *Calculated);
 
-    void getOptimized(const int i, double& lat, double& lon) const;
+    inline size_t Count() const {
+      return m_Task.size();
+    }
+
     void UpdateTaskPoint(const int i, TASK_POINT& TskPt ) const;
 
-    inline size_t Count() const {
-        return m_Task.size();
-    }
+protected:
+    GeoPoint  getOptimized(const int i) const;
 
 protected:
     void AddCircle(int TpIndex);
@@ -37,34 +41,9 @@ protected:
     void AddCone(int TpIndex);
     void AddEssCircle(int TpIndex);
 
-    void Grid2LatLon(double N, double E, double& lat, double& lon) const;
-    void LatLon2Grid(double lat, double lon, double& N, double& E) const;
-
-    typedef struct _DTM {
-        double a; // a  Equatorial earth radius
-        double b; // b  Polar earth radius
-        double f; // f= (a-b)/a  Flatenning
-        double esq; // esq = 1-(b*b)/(a*a)  Eccentricity Squared
-        double e; // sqrt(esq)  Eccentricity
-    } DATUM, *PDATUM;
-
-    typedef struct _GRD {
-        double lon0;
-        double lat0;
-        double k0;
-        double false_e;
-        double false_n;
-    } GRID, *PGRID;
-
     typedef std::vector<PGTaskPt*> Task_t;
-
-#ifdef _WGS84
-    static const DATUM m_Datum_WGS84;
-#endif
-    static const DATUM m_Datum_FAI;
-    GRID m_Grid;
-
     Task_t m_Task;
+    std::unique_ptr<TransverseMercator> m_Projection;
 };
 
 #endif /* PGTASKMGR_H_ */
